[WIP] from 8.0 to 9.0 base_optional_quick_create * bump version to 9.0.1.0.0 and make it installable * apply a views/models directory arch into the module * fix xpath anchor on ir_model views (osv_memory to transient) * add README.rst * apply some new api
