--- base_optional_quick_create/AUTHORS.txt
@@ -1,1 +1,0 @@
-Lorenzo Battistini <lorenzo.battistini@agilebg.com>
