+++ base_optional_quick_create/models/__init__.py
@@ -2,4 +2,4 @@
 # © 2013 Agile Business Group sagl (<http://www.agilebg.com>)
 # © 2016 ACSONE SA/NA (<http://acsone.eu>)
 
-from . import models
+from . import ir_model
