+++ base_optional_quick_create/views/model_view.xml
@@ -5,7 +5,7 @@
             <field name="model">ir.model</field>
             <field name="inherit_id" ref="base.view_model_form"></field>
             <field name="arch" type="xml">
-                <field name="osv_memory" position="after">
+                <field name="transient" position="after">
                     <field name="avoid_quick_create"/>
                 </field>
             </field>
